# server/src/openapi.yaml
openapi: 3.0.3
info:
  title: OrgTree API
  description: |
    RESTful API for OrgTree - an organizational directory and visualization platform.

    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.

    ## Rate Limiting
    - Authentication endpoints: 5 requests per 15 minutes per IP
    - Password reset: 10 requests per 15 minutes per IP
    - General API: No specific limits (recommended: implement client-side throttling)

    ## Real-time Updates
    OrgTree uses Socket.IO for real-time synchronization. Connect to the same server
    with your JWT token to receive live updates for organizations you're a member of.
  version: 1.0.0
  contact:
    name: OrgTree Support
  license:
    name: GPL 3.0
  servers:
    - url: http://localhost:3001/api
      description: Development server
    - url: https://orgtree-app.onrender.com/api
      description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login endpoint
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: 'Error description'
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdById:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        role:
          type: string
          enum: [owner, admin, editor, viewer]
          description: Current user's role in this organization
...
    ShareSettings:
      type: object
      properties:
        isPublic:
          type: boolean
        shareToken:
          type: string
          nullable: true
        shareUrl:
          type: string
          format: uri
          nullable: true
    OwnershipTransfer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
        fromUserId:
          type: string
          format: uuid
        fromUserName:
          type: string
        toUserId:
          type: string
          format: uuid
        toUserName:
          type: string
        toUserEmail:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled, expired]
        reason:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Access token required'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Access denied'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Resource not found'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      description: Get all organizations the user has access to
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Organizations]
      summary: Create organization
      description: Create a new organization. The creator becomes the owner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{orgId}:
    parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Organizations]
      summary: Get organization
      description: Get organization details
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Organizations]
      summary: Update organization
      description: Update organization name. Requires editor role or above.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Organizations]
      summary: Delete organization
      description: Delete organization and all its data. Requires owner role.
      responses:
        '204':
          description: Organization deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== SHARING ====================
  /organizations/{orgId}/share:
    parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Sharing]
      summary: Get share settings
      description: Get public sharing configuration. Requires viewer role or above.
      responses:
        '200':
          description: Share settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Sharing]
      summary: Update share settings
      description: Enable/disable public sharing. Requires admin role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isPublic]
              properties:
                isPublic:
                  type: boolean
      responses:
        '200':
          description: Share settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /organizations/{orgId}/share/regenerate:
    parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Sharing]
      summary: Regenerate share token
      description: Generate a new share URL, invalidating the old one. Requires admin role.
      responses:
        '200':
          description: New share token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== DEPARTMENTS ====================
...
  /organizations/{orgId}/ownership/transfer:
    parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [Ownership Transfers]
      summary: Initiate ownership transfer
      description: Start the process of transferring organization ownership. Requires owner role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [toUserId, reason]
              properties:
                toUserId:
                  type: string
                  format: uuid
                reason:
                  type: string
      responses:
        '201':
          description: Transfer initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transfer:
                    $ref: '#/components/schemas/OwnershipTransfer'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /organizations/{orgId}/ownership/transfers:
    parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Ownership Transfers]
      summary: List organization transfers
      description: List all transfers for an organization. Requires admin role or above.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, cancelled, expired]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of transfers
          content:
            application/json:
              schema:
                type: object
                properties:
                  transfers:
                    type: array
                    items:
                      $ref: '#/components/schemas/OwnershipTransfer'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...